html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/window-iframe-messagechannel.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Structured cloning of SharedArrayBuffers into windows using MessageChannel</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#structuredserialize">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-incrementer.js"></script>
<script src="/common/get-host-info.sub.js"></script>

<div id="log"></div>

<script>
promise_test(t => {
  return new Promise(resolve => {
    const iframe = document.createElement("iframe");
    window.onmessage = t.step_func((message) => {
      // data will be a MessagePort
      resolve(testSharingViaIncrementerScript(t, message.data, "window", message.data, "iframe"));
    });
    iframe.src = "resources/incrementer-iframe-messagechannel.html";
    document.body.appendChild(iframe);
  });
}, `postMessaging to a same-origin iframe via MessageChannel allows them to see each others' modifications`);

promise_test(t => {
  return new Promise(resolve => {
    const iframe = document.createElement("iframe");
    window.onmessage = t.step_func((message) => {
      // data will be a MessagePort
      message.data.postMessage(new SharedArrayBuffer(10));
      message.data.onmessage = t.step_func(message => {
        assert_equals(message.data, "messageerror event received");
        resolve();
      });
    });
    iframe.src = get_host_info().HTTPS_REMOTE_ORIGIN + new URL("resources/iframe-messagechannel-site-failure.html", window.location).pathname;
    document.body.appendChild(iframe);
  });
}, "postMessaging to a same-site iframe via MessageChannel should fail");

promise_test(t => {
  return new Promise(resolve => {
    const iframe = document.createElement("iframe");
    window.onmessage = t.step_func(message => {
      // data will be a MessagePort
      message.data.postMessage(new SharedArrayBuffer(10));
      message.data.onmessage = t.step_func(message => {
        assert_equals(message.data, "messageerror event received");
        resolve();
      });
    });
    iframe.src = get_host_info().HTTPS_NOTSAMESITE_ORIGIN + new URL("resources/iframe-messagechannel-failure.html", window.location).pathname;
    document.body.append(iframe);
  });
}, "postMessaging to a cross-site iframe via MessageChannel should fail");

promise_test(async t => {
  const iframe = document.createElement("iframe");
  const path =
    new URL("resources/iframe-messagechannel-complex.html", location).pathname;
  iframe.src = `${get_host_info().HTTPS_NOTSAMESITE_ORIGIN}${path}`;
  document.body.append(iframe);

  const [port1, port2] = await new Promise((resolve) => {
    let port1, port2;
    window.onmessage = t.step_func(e => {
    if (e.data.state === "port1") {
      port1 = e.data.data;
      e.source.postMessage("send port2", "*");
    } else if (e.data.state === "port2") {
      port2 = e.data.data;
      resolve([port1, port2]);
    }
  })});
  port1.postMessage(new SharedArrayBuffer(10));
  port2.onmessageerror = t.unreached_func();
  await new Promise((resolve) => {
    // Note that onmessage calls start()
    port2.onmessage = t.step_func(e => {
      assert_true(e.data instanceof SharedArrayBuffer);
      assert_equals(e.data.byteLength, 10);
      resolve();
    });
  });
}, "postMessaging with a MessageChannel that's been cross-site should succeed");
</script>
